requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.8.1
cryptography==2.7
fake-useragent==0.1.11
Flask==1.1.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
logger==1.4
MarkupSafe==1.1.1
ntlm-auth==1.3.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
python-nmap==0.6.1
requests==2.22.0
requests-file==1.4.3
requests-ntlm==1.1.0
six==1.12.0
sqlmap==1.3.8
tldextract==2.2.1
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
